requirements.txt
setuptools~=53.0.0
PyYAML~=5.4.1
requests~=2.25.1
sh~=1.14.1